scripts/verify_integration_tests.py
#!/usr/bin/env python3
"""
Simple Integration Test Verification

This script verifies that the integration test setup would work without requiring
a full PostgreSQL installation. It performs basic checks and mock tests.
"""

import sys
import os
from pathlib import Path
from unittest.mock import AsyncMock, MagicMock

# Add project root to path
project_root = Path(__file__).parent.parent
sys.path.insert(0, str(project_root))

try:
    from backend.app.lims.agents.sample_testing import SampleTestingAgent
    print("✓ Sample Testing Agent import successful")
except ImportError as e:
    print(f"✗ Failed to import Sample Testing Agent: {e}")
    sys.exit(1)

try:
    import pytest
    print("✓ pytest available")
except ImportError:
    print("✗ pytest not available")
    sys.exit(1)

try:
    import asyncpg
    print("✓ asyncpg available")
except ImportError:
    print("✗ asyncpg not available - install with: pip install asyncpg")
    sys.exit(1)

try:
    import psutil
    print("✓ psutil available")
except ImportError:
    print("✗ psutil not available - install with: pip install psutil")
    sys.exit(1)

async def test_mock_integration():
    """Test integration test structure with mock database"""
    print("\nTesting integration test structure...")
    
    # Mock database pool
    mock_pool = AsyncMock()
    mock_conn = AsyncMock()
    
    # Mock the context manager
    class MockContextManager:
        def __init__(self, conn):
            self.conn = conn
        
        async def __aenter__(self):
            return self.conn
        
        async def __aexit__(self, exc_type, exc_val, exc_tb):
            pass
    
    mock_pool.acquire = MagicMock(return_value=MockContextManager(mock_conn))
    
    # Create agent
    agent = SampleTestingAgent(mock_pool)
    
    # Test basic operations
    mock_conn.fetch.return_value = [{'test_id': 1}]
    mock_conn.fetchrow.return_value = {'instrument_id': 1}
    
    await agent.process_scheduled_tests()
    
    # Verify mocks were called
    assert mock_conn.fetch.called
    assert mock_conn.fetchrow.called
    assert mock_conn.execute.called
    
    print("✓ Mock integration test structure works")
    
    # Test complete workflow simulation
    mock_conn.reset_mock()
    
    # Complete test workflow
    await agent.complete_test(1, 'pass')
    assert mock_conn.execute.called
    
    mock_conn.reset_mock()
    await agent.transition_to_qc()
    assert mock_conn.execute.called
    
    print("✓ Complete workflow simulation works")

def check_integration_test_files():
    """Check if integration test files exist and are properly structured"""
    print("\nChecking integration test files...")
    
    integration_test_file = project_root / 'tests' / 'test_sample_testing_agent_integration.py'
    if integration_test_file.exists():
        print("✓ Integration test file exists")
        
        # Check file content
        content = integration_test_file.read_text()
        
        checks = [
            ('DatabaseFixture', 'Database fixture class'),
            ('test_complete_workflow_single_test', 'Complete workflow test'),
            ('test_concurrent_agent_operations', 'Concurrent operations test'),
            ('test_high_volume_processing', 'High volume test'),
            ('test_throughput_measurement', 'Performance test'),
        ]
        
        for check, description in checks:
            if check in content:
                print(f"✓ {description} found")
            else:
                print(f"✗ {description} missing")
    else:
        print("✗ Integration test file not found")
        return False
    
    runner_script = project_root / 'scripts' / 'run_integration_tests.py'
    if runner_script.exists():
        print("✓ Integration test runner script exists")
        if os.access(runner_script, os.X_OK):
            print("✓ Runner script is executable")
        else:
            print("! Runner script is not executable (run: chmod +x scripts/run_integration_tests.py)")
    else:
        print("✗ Integration test runner script not found")
        return False
    
    requirements_file = project_root / 'requirements' / 'integration-tests.txt'
    if requirements_file.exists():
        print("✓ Integration test requirements file exists")
    else:
        print("✗ Integration test requirements file not found")
        return False
    
    return True

def main():
    """Main verification function"""
    print("Sample Testing Agent Integration Test Verification")
    print("=" * 60)
    
    # Check basic requirements
    print("Checking basic requirements...")
    
    # Check integration test files
    if not check_integration_test_files():
        print("\n✗ Integration test files are incomplete")
        sys.exit(1)
    
    # Run mock integration test
    import asyncio
    try:
        asyncio.run(test_mock_integration())
    except Exception as e:
        print(f"✗ Mock integration test failed: {e}")
        sys.exit(1)
    
    print("\n" + "=" * 60)
    print("✓ All verification checks passed!")
    print("\nTo run actual integration tests with PostgreSQL:")
    print("1. Install PostgreSQL (version 12+)")
    print("2. Create a test database:")
    print("   createdb alims_test")
    print("3. Run integration tests:")
    print("   python scripts/run_integration_tests.py --setup-db --verbose")
    print("\nOr use Docker:")
    print("   docker run -d --name postgres-test \\")
    print("     -e POSTGRES_USER=postgres \\")
    print("     -e POSTGRES_PASSWORD=postgres \\")
    print("     -e POSTGRES_DB=alims_test \\")
    print("     -p 5432:5432 postgres:14")
    print("   python scripts/run_integration_tests.py --setup-db --verbose")
    print("   docker stop postgres-test && docker rm postgres-test")

if __name__ == '__main__':
    main()
